requirements.txt
APScheduler==3.11.0
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cssselect2==0.8.0
dnspython==2.8.0
email_validator==2.3.0
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.60.0
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.43
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
zopfli==0.2.3.post1
gunicorn==23.0.0
waitress==3.0.2
psycopg2-binary==2.9.9